requirements.txt
asyncio
beautifulsoup4
colorama
discord
python-dotenv
requests
termcolor